requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
albumentations>=1.0.0
opencv-python>=4.5.0
pillow>=8.0.0
scikit-learn>=0.24.0
tqdm>=4.60.0
tensorboard>=2.5.0
PyYAML>=5.4.0